refactor(vtc): split exchange.rs into issue/verify/pending/jwt modules (P2.3)
`credentials/exchange.rs` had grown to 2,585 lines owning four distinct
concerns. Split it into a module directory, re-exported from `exchange/mod.rs`
so every existing `exchange::*` path (credentials/mod.rs, join::retention,
routes::recognise, tests) is byte-for-byte unchanged:

- `issue.rs`   — OID4VCI issuer gate (`verify_oid4vci_proof`, `issue_on_request`,
                 `credential_offer`, `ProvenHolderProof`).
- `verify.rs`  — OID4VP / SD-JWT-VC / W3C-DI / bbs-2023 verifier stack + the
                 `DidVmResolver` key resolver.
- `pending.rs` — persisted single-use pending-offer store (`make_offer`,
                 `redeem`, `sweep_expired_pending`).
- `jwt.rs`     — shared temporal / JWK / JWT-segment helpers, `pub(super)` so the
                 three siblings share one copy.
- `tests.rs`   — the test module, driven through the re-exported public surface.

Pure code-move: no behaviour change, no public-path change. Cross-module helper
calls are the only edits (the jwt helpers + `issue::{credential_offer,
issue_on_request}` are now reached via `super::`). `cargo test -p vtc-service`
green (580 lib + join/recognise/ceremony integration suites); fmt + clippy clean.

`verify.rs` lands at ~900 lines (the SD-JWT + DI-VP + bbs-2023 stack); the
bbs-2023 block is the natural next extraction if it needs to come down further.

Unblocks P2.4 (the DID-VM → DI-proof verifier dedup, whose canonical copy is now
`verify::DidVmResolver`).
